+++ js_api/src/index.ts
@@ -6,6 +6,7 @@ import staking from "./service/staking";
 // import wc from "./service/walletconnect";
 import gov from "./service/gov";
 import parachain from "./service/parachain";
+import assets from "./service/assets";
 import { genLinks } from "./utils/config/config";
 
 // console.log will send message to MsgChannel to App
@@ -60,6 +61,7 @@ const settings = {
 (<any>window).staking = staking;
 (<any>window).gov = gov;
 (<any>window).parachain = parachain;
+(<any>window).assets = assets;
 
 // walletConnect supporting is not ready.
 // (<any>window).walletConnect = wc;