+++ lib/api/api.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:polkawallet_sdk/api/apiAccount.dart';
+import 'package:polkawallet_sdk/api/apiAssets.dart';
 import 'package:polkawallet_sdk/api/apiGov.dart';
 import 'package:polkawallet_sdk/api/apiKeyring.dart';
 import 'package:polkawallet_sdk/api/apiParachain.dart';
@@ -39,6 +40,7 @@ class PolkawalletApi {
   ApiStaking staking;
   ApiGov gov;
   ApiParachain parachain;
+  ApiAssets assets;
   ApiUOS uos;
   ApiRecovery recovery;
 
@@ -55,6 +57,7 @@ class PolkawalletApi {
     staking = ApiStaking(this, service.staking);
     gov = ApiGov(this, service.gov);
     parachain = ApiParachain(this, service.parachain);
+    assets = ApiAssets(this, service.assets);
     uos = ApiUOS(this, service.uos);
     recovery = ApiRecovery(this, service.recovery);